Various fixes for MacOS builds
* Use default provider value for osx_arm64

As described at https://conda-forge.org/docs/maintainer/conda_forge_yml/#provider

* Automatically detect correct clang version in build.sh

* Force one modulemap for Darwin

* Remove checks for SDK versions in cling

* Avoid link TPython against libPython on MacOS

Use the correct CMake target `Python::Module` in the TPython CMakeLists.txt. This will create a libROOTTPython shared library that does not link on MacOS. In a Python environment, TPython will work, but it will not work from C++. This can only work if the library is linked against libpython. To represent the same situation as before this commit, the patch is applied only on MacOS. In fact currently the conda build links against libpython on Linux but on MacOS.